+++ docs/release-notes.md
@@ -10,6 +10,7 @@ hide:
 
 ### Fixed
 - Fix m2m returning through model when using create.
+- Fix incorrect use of delete and save in relations. They should use the raw versions as they are not userfacing.
 
 ## 0.35.0